requirements.txt
pyttsx3
speechrecognition
requests
beautifulsoup4
openai
flask
python-dotenv